# docs/swagger.yaml
definitions:
  app.Pager:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_rows:
        type: integer
    type: object
  errcode.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      msg:
        type: string
    type: object
  model.Article:
    properties:
      content:
        type: string
      cover_image_url:
        type: string
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      desc:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      state:
        type: integer
      title:
        type: string
    type: object
  model.ArticleSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
        type: object
    type: object
  model.Tag:
    properties:
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      name:
        type: string
      state:
        type: integer
    type: object
  model.TagSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
        type: object
    type: object
info:
  contact: {}
  description: Go 语言编程之旅：一起用 Go 做项目
  license: {}
  termsOfService: https://github.com/xudarling/go_tour_http_blog
  title: 博客系统
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: 文章名称
        in: query
        maxLength: 100
        name: title
        type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.ArticleSwagger'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个文章
    post:
      parameters:
      - description: 文章名称
        in: body
        maxLength: 100
        minLength: 3
        name: title
        required: true
        schema:
          type: string
      - description: 文章描述
        in: body
        maxLength: 100
        minLength: 3
        name: desc
        required: true
        schema:
          type: string
      - description: 文章内容
        in: body
        maxLength: 100
        minLength: 3
        name: content
        required: true
        schema:
          type: string
      - description: 文章图片地址
        in: body
        maxLength: 100
        minLength: 3
        name: coverImageUrl
        required: true
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: 创建者
        in: body
        maxLength: 100
        minLength: 3
        name: created_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新增文章
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除文章
    get:
      parameters:
      - description: 文章名称
        in: query
        maxLength: 100
        name: title
        type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取单个文章
    put:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文章名称
        in: body
        maxLength: 100
        minLength: 3
        name: title
        schema:
          type: string
      - description: 文章描述
        in: body
        maxLength: 100
        minLength: 3
        name: desc
        schema:
          type: string
      - description: 文章内容
        in: body
        maxLength: 100
        minLength: 3
        name: content
        schema:
          type: string
      - description: 文章图片地址
        in: body
        maxLength: 100
        minLength: 3
        name: coverImageUrl
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: 修改者
        in: body
        maxLength: 100
        minLength: 3
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Tag'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新文章
  /api/v1/tags:
    get:
      parameters:
      - description: 标签名称
        in: query
        maxLength: 100
        name: name
        type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.TagSwagger'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个标签
    post:
      parameters:
      - description: 标签名称
        in: body
        maxLength: 100
        minLength: 3
        name: name
        required: true
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: 创建者
        in: body
        maxLength: 100
        minLength: 3
        name: created_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新增标签
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除标签
    put:
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签名称
        in: body
        maxLength: 100
        minLength: 3
        name: name
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: 修改者
        in: body
        maxLength: 100
        minLength: 3
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Tag'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新标签
swagger: "2.0"